// src/app/models/character.ts
export interface CharacterI {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string,
    url: string
  };
  location: {
    name: string
    url: string
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export interface Info {
  count: number
  pages: number
  next: string
  prev: string
}